Se hicieron los cambios

+++ src/functions/users/user.service.ts
@@ -110,8 +110,10 @@ export default class UserServices {
         console.log(`Service START --> ${this.deleteUser.name}`);
         try {
             const userP = await UserPropertiesDAO.findUserById(userId);
+            if (!userP) throw new ConflictException('VENTAS_MM_COMMON_NOT_FOUND_404', { error: `user ${userId} not exist` });
             await UserPropertiesDAO.deleteUser(userP);
             const user = await UserDAO.findUserById(userId);
+            if (!user) throw new ConflictException('VENTAS_MM_COMMON_NOT_FOUND_404', { error: `user ${userId} not exist` });
             await UserDAO.deleteUser(user);
             console.log(`Service END --> ${this.deleteUser.name}`); 
         } catch (error) {
